강제 업데이트 화면이 전체를 덮지 않는 문제
하단 탭이 보이고 있어서 수정함
